+++ index.html
@@ -15,16 +15,15 @@
     <body>
     
     
-             </div>
             <div class="row">
                 <div class="col">
                     Graph vertices:
-                    <input id="inputGraphVertices" class="form-control" size="1" value="8" />
+                    <input id="inputGraphVertices" class="form-control" size="1" value="8" type="number" min="1"  />
                     <br> Edge probability:
-                    <input id="inputGraphProb" size="1" class="form-control" value="0.5" /> Vertex radius:
-                    <input id="inputGraphRadius" size="1" class="form-control" value="0.18" />
+                    <input id="inputGraphProb" size="1" class="form-control" value="0.5" type="number" min="0" max="1" step="0.05"/> Vertex radius:
+                    <input id="inputGraphRadius" size="1" class="form-control" value="0.18" type="number" min="0"/>
                     <br> Edge length:
-                    <input id="inputGraphLength" size="1" class="form-control" value="1" />
+                    <input id="inputGraphLength" size="1" class="form-control" value="1" type="number" min="0"  />
                     <br>
                     <button id="buttonSpawnGraph" class="btn btn-success">Spawn Graph</button>
                     <button id="buttonDestroyGraph" class="btn btn-danger">Destroy Graph</button>
